v3.1.0: New/improved DOS/FAT file naming scheme Now uses "VOLxx...PO" filenames for FAT, and only considers the first 5 bytes (and the file extension).
So you can now use longer filenames on the SD cards, to keep files organized, e.g.:
	VOL00_TOTALREPLAY.PO
	VOL01_InstantReplay.PO
	...
	VOL0F_FooBar.PO
etc

The old "BLKDEVxx.PO" filename scheme is also still supported to keep
backward compatibility with old SD cards. However, you cannot mix the
naming schemes (not on a single SD card, nor on different SD cards plugged
at the same time).

Also, the old "BLKDEVxx.PO" filenames have to match exactly (just as before):
you cannot use longer filenames, as is now supported with the new naming scheme.
This has to do with how the "long filename" (LFN) support was implemented to fit
the 8.3 FAT filesystem: as soon as a filename has more than 8 characters, the
last two bytes get cut off and are replaced with "~1" in the 8.3 FAT directory
entry) Hence, we switched to a new and shorter "VOLxx" scheme (so the simple
8.3 filenames still conain the important volume number information).