x44: First revision for the Atmega 644 (and 1284?)
This is a straight derivative version with most small parts converted to
SMT. Also all using footprints and references from JLPCB, so as of
today, you can order a fully built card; it'll just be missing the
Atmega, which is not in stock, but is easy to add...

I still used the 'hand soldering' footprints from Kicad so it should be
easy to modify and rework.

Oh also, I simplified some things, re-layout the schemastic on an A3
sheet, and replaced diodes with a LDO with good filtering, that should
improve SD card compatibility.

I removed some test headers, re-layouted all the connectors at the top,
and moved the LEDs at the front.

To amuse myself I tried to layout the board in a 'classic'
vertical/horizontal pattern of traces, so it's not 'cheap' on VIAs, but
it looks cool! :-)

Signed-off-by: Michel Pollet <buserror@gmail.com>